Move fixtures into proper places in client tests
